utils as macros
